libwcs 1.1
imrot added to rotate and mirror FITS or IRAF images, to deal with
Mt. Hopkins CCD images. Subroutines were added to perform the rotation
and to move pixels between two image arrays, which do not have to be
of the same type.  i2f was added to convert IRAF image files to FITS
image files, making use of subroutines which already existed.  imtab
was added to list stars from a tab table catalog which are found in an
image.  imsize was added to print the center and size of an image in
one line.  imhead was added to print the headers of FITS and IRAF
images.